solution with components
